+++ app/src/main/java/models/Report.kt
@@ -3,8 +3,8 @@ package models
 data class Report(
     val uid: String?,
     val drinkable: Float?,
-    val algae:Double?,
-    val dirty:Double?,
+    val algae:Float?,
+    val dirty:Float?,
     val date: String?,
     val lat: Double?,
     val lon: Double?
